TYB.MLService/requirements.txt
# TYB MLService - Python Bağımlılıkları

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Scheduling
apscheduler==3.10.4

# Machine Learning (pre-built wheels)
numpy==1.26.4
scipy==1.11.4
joblib==1.3.2
scikit-learn==1.4.1.post1

# ONNX Runtime (optional)
onnxruntime==1.17.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging
python-json-logger==2.0.7





